hyperliquid-signing-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
hyperliquid-python-sdk
web3>=6.0.0
python-dotenv>=1.0.0

